관Feat: 리자í 총매ì¶Refactor 수정, 관리자통계 단일책임원칙으로 수정
