+++ src/main/java/com/github/project3/service/admin/AuthService.java
@@ -48,53 +48,5 @@ public class AuthService {
                 .map(AdminEntity::getSales)
                 .orElse(0); // 매출액이 없는 경우 0 반환
     }
-    // CreateAt 기준 Count수 / StartDate기준 전체 매출액
-    public <T extends CreatedAtRepository> long countEntityCreatedBetween(T repository, LocalDateTime start, LocalDateTime end) {
-        return repository.countByCreatedAtBetween(start, end);
-    }
-    public Long sumTotalPriceByStartDateBetween(BookRepository repository, LocalDateTime start, LocalDateTime end){
-        return repository.sumTotalPriceByStartDateBetween(start, end);
-    }
-    // + 기간별 데이터(1일, 1주일, 1달, 전체)
-    public long getLastDayCount(CreatedAtRepository createdAtRepository){
-        LocalDateTime now = LocalDateTime.now();
-        LocalDateTime oneDayAgo = now.minusDays(1);
-        return countEntityCreatedBetween(createdAtRepository, oneDayAgo, now);
-    }
-    public long getLastWeekCount(CreatedAtRepository createdAtRepository){
-        LocalDateTime now = LocalDateTime.now();
-        LocalDateTime oneWeekAgo = now.minusWeeks(1);
-        return countEntityCreatedBetween(createdAtRepository, oneWeekAgo, now);
-    }
-    public long getLastMonthCount(CreatedAtRepository createdAtRepository){
-        LocalDateTime now = LocalDateTime.now();
-        LocalDateTime oneMonthAgo = now.minusMonths(1);
-        return countEntityCreatedBetween(createdAtRepository, oneMonthAgo, now);
-    }
-    public long getLastDaySales(BookRepository bookRepository){
-        LocalDateTime now = LocalDateTime.now();
-        LocalDateTime oneDayAgo = now.minusDays(1);
-        return sumTotalPriceByStartDateBetween(bookRepository, oneDayAgo, now);
-    }
-    public long getLastWeekSales(BookRepository bookRepository){
-        LocalDateTime now = LocalDateTime.now();
-        LocalDateTime oneWeekAgo = now.minusWeeks(1);
-        return sumTotalPriceByStartDateBetween(bookRepository, oneWeekAgo, now);
-    }
-    public long getLastMonthSales(BookRepository bookRepository){
-        LocalDateTime now = LocalDateTime.now();
-        LocalDateTime oneMonthAgo = now.minusMonths(1);
-        return sumTotalPriceByStartDateBetween(bookRepository, oneMonthAgo, now);
-    }
-
-    public long getTotalUserCount(UserRepository userRepository){
-        return userRepository.count();
-    }
-    public long getTotalBookCount(BookRepository bookRepository){
-        return bookRepository.count();
-    }
-    public long getTotalSales(BookRepository bookRepository){
-        return bookRepository.sumTotalPrice();
-    }
 
 }
